server/requirements.txt
fastapi
uvicorn[standard]
pydantic
motor
pymongo
python-dotenv
python-jose
passlib[bcrypt]